# auth/src/main/kotlin/buyticket/auth/dto/Request.kt
package buyticket.auth.dto

data class LoginDto(
    val email: String,
    val password: String,
)

data class RegisterDto(
    val nickName: String,
    val password: String,
    val email: String,
)

data class TokenDto(
    val token: String,
)